+++ packages/colony-js-adapter/index.js
@@ -3,7 +3,11 @@
 export type { Adapter as IAdapter } from './interface/Adapter';
 export type { Block } from './interface/Block';
 export type { Event } from './interface/Event';
-export type { EventHandler, EventHandlers } from './interface/EventHandlers';
+export type {
+  EventHandler,
+  EventHandlers,
+  EventCallback,
+} from './interface/EventHandlers';
 export type { Contract as IContract } from './interface/Contract';
 export type { Provider as IProvider } from './interface/Provider';
 export type { Signature } from './interface/Signature';