+++ packages/colony-js-contract-client/src/__tests__/ContractMethod.js
@@ -5,190 +5,26 @@ import createSandbox from 'jest-sandbox';
 
 import ContractClient from '../classes/ContractClient';
 import ContractMethod from '../classes/ContractMethod';
-
-import {
-  validateValue,
-  convertInputValue,
-  convertOutputValue,
-} from '../modules/paramTypes';
-
-jest.mock('../modules/paramTypes', () => ({
-  validateValue: jest.fn().mockImplementation(() => true),
-  convertOutputValue: jest
-    .fn()
-    .mockImplementation(value => `converted output: ${value}`),
-  convertInputValue: jest
-    .fn()
-    .mockImplementation(value => `converted input: ${value}`),
-}));
+import * as types from '../modules/paramTypes';
 
 describe('ContractMethod', () => {
   const sandbox = createSandbox();
   const client = new ContractClient({});
+
   sandbox
     .spyOn(client, 'createTransactionData')
     .mockImplementation(() => 'the tx data');
 
   beforeEach(() => {
     sandbox.clear();
-    convertInputValue.mockClear();
-    convertOutputValue.mockClear();
-    validateValue.mockClear();
-  });
 
-  test('Method with no input params validates correctly', () => {
-    const method = new ContractMethod({ client, functionName: 'myFunction' });
-    expect(method.validate()).toBe(true);
-  });
-
-  test('Method with input params validates correctly', () => {
-    const input = [
-      ['taskId', 'number'],
-      ['potId', 'number'],
-      ['domainId', 'number', 1],
-    ];
-    const inputValues = { taskId: 6, potId: 420 };
-
-    const method = new ContractMethod({
-      client,
-      input,
-      functionName: 'myFunction',
-      name: 'myMethodName',
-    });
-    sandbox.spyOn(method, '_validateValue');
-
-    expect(method.validate(inputValues)).toBe(true);
-    expect(method._validateValue).toHaveBeenCalledWith(6, input[0]);
-    expect(method._validateValue).toHaveBeenCalledWith(420, input[1]);
-
-    // Missing parameters/wrong type
-    [undefined, null, [], 'a', 6].forEach(wrongType => {
-      expect(() => {
-        method.validate(wrongType);
-      }).toThrowError('Expected parameters as an object');
-    });
-
-    // No parameters
-    expect(() => {
-      method.validate({});
-    }).toThrowError('Missing parameters');
-
-    // Missing parameter
-    expect(() => {
-      method.validate({ taskId: 6 });
-    }).toThrowError('Missing parameters: "potId"');
-
-    // Wine parameters
-    expect(() => {
-      method.validate({
-        caskId: 6,
-        côteId: 420,
-        domaineId: 'Domaine de la Romanee-Conti',
-      });
-    }).toThrowError('Unexpected parameters');
-
-    // Extra parameter
-    expect(() => {
-      method.validate({
-        taskId: 6,
-        domainId: 1,
-        potId: 420,
-        somethingElse: 2,
-      });
-    }).toThrowError('Unexpected parameters: "somethingElse"');
-
-    // Extra parameter, without the parameter that has a default value
-    expect(() => {
-      method.validate({ taskId: 6, potId: 420, somethingElse: 1 });
-    }).toThrowError('Unexpected parameters: "somethingElse"');
-
-    // Wrong type
-    validateValue.mockImplementationOnce(() => false);
-    expect(() => {
-      method.validate({ taskId: 'six', potId: 420 });
-    }).toThrowError('Parameter "taskId" expected a value of type "number"');
-  });
-
-  test('Valid values are reported as valid', () => {
-    const input = [['id', 'number']];
-    const method = new ContractMethod({
-      client,
-      input,
-      functionName: 'myFunction',
-      name: 'myMethodName',
-    });
-    validateValue.mockImplementationOnce(() => true);
-    expect(method._validateValue(1, ['id', 'number'])).toBe(true);
-    expect(validateValue).toHaveBeenCalledWith(1, 'number');
-  });
-
-  test('Invalid values are reported as invalid, with reasons', () => {
-    const input = [['id', 'number']];
-    const method = new ContractMethod({
-      client,
-      input,
-      functionName: 'myFunction',
-      name: 'myMethodName',
-    });
-
-    // Invalid value
-    validateValue.mockImplementationOnce(() => false);
-    expect(() => {
-      expect(method._validateValue('abc', ['id', 'number']));
-    }).toThrowError('Parameter "id" expected a value of type "number"');
-    expect(validateValue).toHaveBeenCalledWith('abc', 'number');
-
-    // Invalid value with reasons (caught validation errors)
-    validateValue.mockImplementationOnce(() => {
-      throw new Error('The reason this validation failed');
-    });
-    expect(() => {
-      expect(method._validateValue('abc', ['id', 'number']));
-    }).toThrowError(
-      'Parameter "id" expected a value of type ' +
-        '"number" (The reason this validation failed)',
-    );
-
-    // Validation error messages contain the method name
-    validateValue.mockImplementationOnce(() => false);
-    expect(() => {
-      expect(method._validateValue('abc', 'number', 'id'));
-    }).toThrowError('Validation failed for myMethod');
-  });
-
-  test('Valid default values are reported as valid', () => {
-    validateValue.mockImplementationOnce(() => true);
-    const input = [['id', 'number']];
-    const method = new ContractMethod({
-      client,
-      input,
-      name: 'myMethodName',
-    });
-    expect(method._validateValue(undefined, ['id', 'number', 1])).toBe(true);
-    expect(validateValue).toHaveBeenCalledWith(1, 'number');
-  });
-
-  test('Invalid default values are reported as invalid', () => {
-    const input = [['id', 'number']];
-    const method = new ContractMethod({
-      client,
-      input,
-      name: 'myMethodName',
-    });
-    validateValue.mockImplementationOnce(() => false);
-    expect(() => {
-      expect(
-        method._validateValue(undefined, [
-          'id',
-          'number',
-          'a bad default value',
-        ]),
-      );
-    }).toThrowError(
-      'Validation failed for myMethodName: Parameter "id" expected a ' +
-        'value of type "number"',
-    );
-    expect(validateValue).toHaveBeenCalledWith('a bad default value', 'number');
+    sandbox
+      .spyOn(types, 'convertInputValue')
+      .mockImplementation(value => `converted input: ${value}`);
+    sandbox
+      .spyOn(types, 'convertOutputValue')
+      .mockImplementation(value => `converted output: ${value}`);
+    sandbox.spyOn(types, 'validateValueType').mockImplementation(() => true);
   });
 
   test('Method arguments are processed from input parameters', () => {
@@ -200,11 +36,11 @@ describe('ContractMethod', () => {
       input,
       functionName: 'myFunction',
     });
-    sandbox.spyOn(method, '_parseInputValues').mockImplementation(() => [1]);
+
+    sandbox.spyOn(method, 'convertInputValues').mockImplementation(() => [1]);
 
     expect(method._getMethodArgs(inputValues)).toEqual([1]);
-
-    expect(method._parseInputValues).toHaveBeenCalledWith(inputValues);
+    expect(method.convertInputValues).toHaveBeenCalledWith(inputValues);
   });
 
   test('Method without input defined gets empty arguments', () => {
@@ -235,6 +71,7 @@ describe('ContractMethod', () => {
       input,
       functionName: 'myFunction',
     });
+
     sandbox.spyOn(method, 'validate');
     sandbox.spyOn(method, '_getMethodArgs');
 
@@ -257,10 +94,16 @@ describe('ContractMethod', () => {
       functionName: 'myFunction',
     });
 
-    const returnValues = method._getOutputValues(callResult);
+    const returnValues = method.convertOutputValues(callResult);
 
-    expect(convertOutputValue).toHaveBeenCalledWith(callResult[0], 'string');
-    expect(convertOutputValue).toHaveBeenCalledWith(callResult[1], 'address');
+    expect(types.convertOutputValue).toHaveBeenCalledWith(
+      callResult[0],
+      'string',
+    );
+    expect(types.convertOutputValue).toHaveBeenCalledWith(
+      callResult[1],
+      'address',
+    );
 
     expect(returnValues).toEqual({
       name: 'converted output: Vitalik',
@@ -287,7 +130,7 @@ describe('ContractMethod', () => {
   });
 
   test('Input values fall back to default values if not provided', () => {
-    convertInputValue.mockImplementation(value => value);
+    types.convertInputValue.mockImplementation(value => value);
 
     const specificationHash = 'QmcNbGg6EVfFn2Z1QxWauR9XY9KhnEcyb5DUXCXHi8pwMJ';
     const domainId = 1;
@@ -303,14 +146,14 @@ describe('ContractMethod', () => {
       functionName: 'myFunction',
     });
 
-    expect(method._parseInputValues(inputValues)).toEqual([
+    expect(method.convertInputValues(inputValues)).toEqual([
       specificationHash,
       domainId,
     ]);
-    expect(convertInputValue).toHaveBeenCalledWith(
+    expect(types.convertInputValue).toHaveBeenCalledWith(
       specificationHash,
       'ipfsHash',
     );
-    expect(convertInputValue).toHaveBeenCalledWith(domainId, 'number');
+    expect(types.convertInputValue).toHaveBeenCalledWith(domainId, 'number');
   });
 });