+++ packages/colony-js-contract-client/src/classes/ContractMethod.js
@@ -4,14 +4,13 @@
 import isPlainObject from 'lodash.isplainobject';
 import { makeAssert } from '@colony/colony-js-utils';
 
-import type { ContractMethodArgs, Params, Param } from '../flowtypes';
 import ContractClient from './ContractClient';
 import {
-  validateValue,
-  convertInputValue,
-  convertOutputValue,
-} from '../modules/paramTypes';
-import { areParamPairsEmpty, isInputEmpty } from '../modules/inputValidation';
+  convertInputValues,
+  convertOutputValues,
+} from '../modules/paramConversion';
+import { validateParams } from '../modules/paramValidation';
+import type { ContractMethodArgs, Params } from '../flowtypes';
 
 /**
  * Abstract class for interacting with contract methods.
@@ -44,48 +43,6 @@ export default class ContractMethod<
   }
 
   /**
-   * Given the result of a contract method call, and the input values used to
-   * make the call, transform these with the expected output parameters in
-   * order to get named output values as the method's `OutputValues`
-   * The inputValues are included for function overloading.
-   */
-  // eslint-disable-next-line no-unused-vars
-  _getOutputValues(callResult: any, inputValues?: InputValues): OutputValues {
-    // It may be a single value or an array; treat it as an array
-    const values = [].concat(callResult);
-
-    // Clean values according to their type and coalesce into an object
-    return this.output && this.output.length
-      ? // $FlowFixMe
-        this.output
-          .map(([name, type], index) => [
-            name,
-            convertOutputValue(values[index], type),
-          ])
-          .reduce(
-            (acc, [name, value]) => Object.assign(acc, { [name]: value }),
-            {},
-          )
-      : callResult;
-  }
-
-  /**
-   * Given input values, map them against the method's expected parameters,
-   * with the appropriate conversion for each type.
-   * Fall back to default values for each parameter.
-   */
-  _parseInputValues(inputValues: InputValues) {
-    return this.input.map(([paramName, paramType, defaultValue]) =>
-      convertInputValue(
-        Object.hasOwnProperty.call(inputValues, paramName)
-          ? inputValues[paramName]
-          : defaultValue,
-        paramType,
-      ),
-    );
-  }
-
-  /**
    * Given named input values, transform these with the expected parameters
    * in order to get an array of arguments expected by the contract function.
    */
@@ -95,7 +52,7 @@ export default class ContractMethod<
     if (inputValues == null) return args;
 
     if (this.input && this.input.length) {
-      args = this._parseInputValues(inputValues);
+      args = this.convertInputValues(inputValues);
     } else if (
       isPlainObject(inputValues) &&
       Object.getOwnPropertyNames(inputValues).length
@@ -118,69 +75,47 @@ export default class ContractMethod<
   }
 
   /**
+   * Given input values, validate them
+   */
+  validate(inputValues?: any) {
+    return validateParams(inputValues, this.input, this.assertValid);
+  }
+
+  /**
+   * Given input values, map them against the method's expected parameters,
+   * with the appropriate conversion for each type.
+   * Fall back to default values for each parameter.
+   */
+  convertInputValues(inputValues: InputValues): Array<any> {
+    return convertInputValues(inputValues, this.input);
+  }
+
+  /**
+   * Given the result of a contract method call, transform these with the
+   * expected output parameters in order to get named output values as the
+   * method's `OutputValues`.
+   */
+  convertOutputValues(
+    callResult: any,
+    // eslint-disable-next-line no-unused-vars
+    inputValues?: InputValues,
+  ): OutputValues {
+    const values = [].concat(callResult);
+
+    const parsedResult = this.output.reduce(
+      (acc, [name], index) => Object.assign(acc, { [name]: values[index] }),
+      {},
+    );
+
+    return convertOutputValues(parsedResult, this.output);
+  }
+
+  /**
    * Given input values, validate them and return parsed method args.
    */
   getValidatedArgs(inputValues?: any) {
     this.validate(inputValues);
+
     return this._getMethodArgs(inputValues);
-  }
-
-  /**
-   * Given parameters (as an object with named parameters, but potentially any
-   * kind of invalid input), validate each parameter against the expected type
-   * for this method, throwing vaidation errors or returning true.
-   */
-  validate(input?: any): boolean {
-    if (areParamPairsEmpty(this.input) && isInputEmpty(input)) return true;
-
-    this.assertValid(isPlainObject(input), 'Expected parameters as an object');
-
-    const inputValues = Object.assign({}, input);
-
-    const paramNames = this.input.map(([name]) => name);
-    const extraParams = Object.keys(inputValues).filter(
-      name => !paramNames.includes(name),
-    );
-    this.assertValid(
-      extraParams.length === 0,
-      `Unexpected parameters: "${extraParams.join(', ')}"`,
-    );
-
-    // Either the parameter name should exist in the inputValues,
-    // or the parameter should have a default value.
-    const missingParams = this.input.filter(
-      param =>
-        !(
-          Object.hasOwnProperty.call(inputValues, param[0]) ||
-          param.length === 3
-        ),
-    );
-    this.assertValid(
-      missingParams.length === 0,
-      `Missing parameters: "${missingParams.map(([name]) => name).join(', ')}"`,
-    );
-
-    return this.input.every(param =>
-      this._validateValue(inputValues[param[0]], param),
-    );
-  }
-
-  _validateValue(value: any, [name, type, defaultValue]: Param) {
-    let reason;
-    let isValid = false;
-    try {
-      isValid = validateValue(
-        typeof value !== 'undefined' ? value : defaultValue,
-        type,
-      );
-    } catch (error) {
-      reason = error.message || error.toString();
-    }
-    return this.assertValid(
-      Boolean(isValid),
-      `Parameter "${name}" expected a value of type "${type}"${
-        reason ? ` (${reason})` : ''
-      }`,
-    );
   }
 }